Splitting tests up so machinery (plots) and ffmpeg stuff are separated out
